# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API",
        "description": "Petship API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://api.petship.pt",
            "description": "Petship API"
        }
    ],
    "paths": {
...
        "/api/enviarmensagem": {
            "post": {
                "tags": [
                    "Mensagens"
                ],
                "summary": "Enviar mensagem a um utilizador",
                "description": "Rota para enviar mensagem a um utilizador, o utilizador tem de estar logado. Se a mensagem for enviada com sucesso retorna o status 200",
                "operationId": "f2a2fb60156538837b0fe5235594b71b",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_recebe",
                                    "mensagem"
                                ],
                                "properties": {
                                    "id_recebe": {
                                        "example": "1"
                                    },
                                    "mensagem": {
                                        "example": "Olá, tudo bem?"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/mensagens/{id_recebe}": {
            "get": {
                "tags": [
                    "Mensagens"
                ],
                "summary": "Ler conversa com um utilizador",
                "description": "Rota para ler a conversa com um utilizador, o utilizador tem de estar logado. Se a conversa for lida com sucesso retorna o status 200",
                "operationId": "906b8fc05b0b2196bca1e128ff4b1fab",
                "parameters": [
                    {
                        "name": "id_recebe",
                        "in": "path",
                        "description": "Id do utilizador com quem quer ler a conversa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/conversasativas": {
            "get": {
                "tags": [
                    "Mensagens"
                ],
                "summary": "Obter conversas ativas",
                "description": "Rota para obter todas as conversas que se tem ativas (com mensagens enviadas ou recebidas). O utilizador tem de estar logado. Se as conversas forem obtidas com sucesso retorna o status 200",
                "operationId": "4b48462fa1bfb72b575e29775a59315a",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "token": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}